server/web/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gevent==20.9.0
gunicorn==20.0.4
pika==1.1.0
python-magic==0.4.18